refactor: 로그아웃 버튼 분리
