+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { useContext, useState } from 'react';
-import { Switch, Route, Link, NavLink, Redirect, useHistory } from 'react-router-dom';
+import { Switch, Route, Link, NavLink, Redirect } from 'react-router-dom';
 
 import Header from './components/Header/Header';
 import NavBar from './components/NavBar/NavBar';
@@ -10,29 +10,18 @@ import { LoginPage, SignupPage, StationPage, LinePage, SectionPage, MapPage } fr
 
 import { ThemeContext } from './contexts/ThemeContextProvider';
 import { UserContext } from './contexts/UserContextProvider';
-import { SnackBarContext } from './contexts/SnackBarProvider';
 
 import PATH from './constants/path';
 import PALETTE from './constants/palette';
-import { SUCCESS_MESSAGE } from './constants/messages';
 
 import { ReactComponent as TextLogo } from './assets/images/ww-subway-logo-text.svg';
+import LogoutButton from './components/shared/Button/LogoutButton';
 
 const App = () => {
-  const history = useHistory();
-
-  const addMessage = useContext(SnackBarContext)?.addMessage;
   const userContext = useContext(UserContext);
   const themeColor = useContext(ThemeContext)?.themeColor ?? PALETTE.WHITE;
 
   const [isLoading, setIsLoading] = useState<boolean>(false);
-
-  const onLogout = async () => {
-    await userContext?.setIsLoggedIn(false);
-
-    addMessage?.(SUCCESS_MESSAGE.LOGOUT);
-    history.push(PATH.LOGIN);
-  };
 
   return (
     <>
@@ -46,13 +35,7 @@ const App = () => {
           <NavLink to={PATH.SECTIONS}>구간관리</NavLink>
           <NavLink to={PATH.MAP}>전체보기</NavLink>
           {userContext?.isLoggedIn ? (
-            <button
-              onClick={() => {
-                onLogout();
-              }}
-            >
-              로그아웃
-            </button>
+            <LogoutButton />
           ) : (
             <>
               <NavLink to={PATH.LOGIN}>로그인</NavLink>
